@@ -45,7 +45,7 @@
     "lodash.sample": "^4.2.1",
     "lodash.shuffle": "^4.2.0",
     "lodash.values": "^4.3.0",
-    "redis-commands": "^1.2.0",
+    "redis-commands": "^1.3.5",
     "redis-errors": "^1.2.0",
     "redis-parser": "^3.0.0",
     "standard-as-callback": "^1.0.0"
@@ -65,7 +65,7 @@
     "typescript": "^3.1.1"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=6"
   },
   "config": {
     "commitizen": {
